@@ -1,6 +1,6 @@
 def versionMajor = 3
 def versionMinor = 3
-def versionPatch = 1
+def versionPatch = 3
 
 buildscript {
     repositories {
@@ -20,7 +20,7 @@ apply plugin: 'io.fabric'
 
 repositories {
     mavenCentral()
-
+    maven { url 'https://maven.google.com/' }
     maven { url 'https://maven.fabric.io/public' }
     maven { url 'http://devrepo.kakao.com:8088/nexus/content/groups/public/' }
 
@@ -37,12 +37,12 @@ android {
 //            storePassword 'tifmdgnp1'
         }
     }
-    compileSdkVersion 27
+    compileSdkVersion 30
     useLibrary 'org.apache.http.legacy'
     defaultConfig {
         applicationId "univ.sm"
         minSdkVersion 16
-        targetSdkVersion 26
+        targetSdkVersion 30
         versionCode versionMajor * 10000 + versionMinor * 100 + versionPatch
         versionName "${versionMajor}.${versionMinor}.${versionPatch}"
         multiDexEnabled true
@@ -62,6 +62,7 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+    buildToolsVersion = '29.0.2'
 }
 
 dependencies {
@@ -80,12 +81,12 @@ dependencies {
     /** ***************************************************************/
     /** ***************************************************************/
     //compile 'com.android.support:support-v4:27.1.0'
-    implementation 'com.android.support:appcompat-v7:27.0.1'
-    implementation 'com.android.support:design:27.0.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
     // RecyclerView
-    implementation 'com.android.support:recyclerview-v7:27.0.1'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
     // CardView
-    implementation 'com.android.support:cardview-v7:27.0.1'
+    implementation 'androidx.cardview:cardview:1.0.0'
 
     // Glide
     implementation 'com.github.bumptech.glide:glide:3.7.0'
@@ -112,8 +113,8 @@ dependencies {
 
 
     /********************** BUTTER NIFE **********************/
-    implementation 'com.jakewharton:butterknife:8.8.1'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
+    implementation 'com.jakewharton:butterknife:10.0.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.0.0'
     /********************** BUTTER NIFE **********************/
 
 
