@@ -1,33 +1,18 @@
 package univ.sm.view.entry;
 
-import android.content.Context;
-import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+
 import android.view.KeyEvent;
-import android.view.View;
-import android.view.ViewTreeObserver;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
-import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
+import android.webkit.WebViewClient;
 
-import com.github.aakira.expandablelayout.ExpandableLayout;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdView;
 
-import java.util.ArrayList;
-
 import univ.sm.R;
-import univ.sm.StaticData;
-import univ.sm.model.Const;
-import univ.sm.model.Utility;
-import univ.sm.model.shuttle.Shuttle;
 import univ.sm.view.CommonView;
-import univ.sm.view.entry.adapter.EntryRecyclerAdapter;
 
 /**
  * Created by heesun on 2016-12-13.
@@ -46,7 +31,8 @@ public class SchEntryWebView extends CommonView{
 
 
         webView = (WebView) findViewById(R.id.webview);
-        webView.loadUrl("http://sm.choi.link");
+        webView.setWebViewClient(new WebViewClient());
+        webView.loadUrl("https://sm.choi.link");
         WebSettings webSettings = webView.getSettings();
         webSettings.setJavaScriptEnabled(true);
 
