@@ -2,8 +2,8 @@ package univ.sm.view.question;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewPager;
+import androidx.annotation.Nullable;
+import androidx.viewpager.widget.ViewPager;
 import android.widget.RadioGroup;
 
 import com.viewpagerindicator.CirclePageIndicator;
